+++ catatom2osm/__init__.py
@@ -1,2 +1,2 @@
-__version__ = '1.8.0'
+__version__ = '1.8.1'
 from catatom2osm.__main__ import run